core/core/requirements/base.txt
appdirs==1.4.4
appnope==0.1.3
asgiref==3.6.0
asttokens==2.2.1
async-generator==1.10
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
black==21.6b0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
coverage==5.5
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20230430
distlib==0.3.6
dj-database-url==2.0.0
Django==4.2
django-countries==7.5.1
-e git+https://github.com/jazzband/django-debug-toolbar.git@75da65ccc17add38d7cab79d69f93db4ae495d31#egg=django_debug_toolbar
django-import-export==3.2.0
django-js-asset==2.0.0
django-money==3.1.0
django-mptt==0.14.0
django-phonenumber-field==7.1.0
django-tinymce==3.6.1
django-versatileimagefield==3.0
et-xmlfile==1.1.0
exceptiongroup==1.1.1
executing==1.2.0
factory-boy==3.2.0
Faker==18.6.0
filelock==3.12.0
flake8==3.9.2
gunicorn==20.1.0
h11==0.14.0
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
ipython==7.25.0
isort==5.9.1
jedi==0.18.2
mailchimp-marketing==3.0.80
MarkupPy==1.14
matplotlib-inline==0.1.6
mccabe==0.6.1
mypy-extensions==1.0.0
odfpy==1.4.1
openpyxl==3.1.2
outcome==1.2.0
packaging==23.1
parso==0.8.3
pathspec==0.11.1
pexpect==4.8.0
phonenumbers==8.13.11
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.5.0
pluggy==1.0.0
prompt-toolkit==3.0.38
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
py-moneyed==2.0
pycodestyle==2.7.0
pyflakes==2.3.1
Pygments==2.15.1
PySocks==1.7.1
pytest==6.2.5
pytest-base-url==2.0.0
pytest-django==4.5.2
pytest-factoryboy==2.5.1
pytest-html==3.2.0
pytest-metadata==2.0.4
pytest-selenium==4.0.0
pytest-variables==2.0.0
python-dateutil==2.8.2
python-decouple==3.8
python-magic==0.4.27
python-magic-bin==0.4.14
pytz==2023.3
PyYAML==6.0
regex==2023.3.23
requests==2.29.0
selenium==4.9.1
sentry-sdk==1.21.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.4
stack-data==0.6.2
tablib==3.4.0
tenacity==6.3.1
toml==0.10.2
tomli==2.0.1
tox==3.23.1
traitlets==5.9.0
trio==0.22.0
trio-websocket==0.10.2
typing_extensions==4.5.0
urllib3==1.26.15
virtualenv==20.23.0
wcwidth==0.2.6
wsproto==1.2.0
xlrd==2.0.1
xlwt==1.3.0
